Rearrange testings
